LayoutTests/platform/mac/accessibility/search-field-cancel-button-expected.txt


This tests that the search field cancel button is exposed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS button.description is 'AXDescription: cancel'
PASS button.role is 'AXRole: AXButton'
PASS button.roleDescription is 'AXRoleDescription: button'
PASS document.getElementById('search').value is 'X'
PASS document.getElementById('search').value is ''
PASS successfullyParsed is true

TEST COMPLETE
